moves main code into src
